hk-tick-collector/requirements.txt
aiosqlite>=0.19.0
PyYAML>=6.0.1
aiohttp>=3.9.0
pandas>=2.0.0
futu>=0.0.1

tzdata>=2024.1
